Refactored the membership pallet
