+++ runtime/src/configs/mod.rs
@@ -2,3 +2,4 @@ pub mod authorship;
 pub mod balances;
 pub mod timestamp;
 pub mod transaction_payment;
+pub mod membership;
